Move OpenUrlModal contents into a component
