+++ src/main.js
@@ -3,7 +3,7 @@ const path = require('path');
 const url = require('url');
 const setupMenus = require('./menus');
 const wrapNavigation = require('./wrapNavigation');
-const {createOpenUrlModal} = require('./openUrlMain');
+const {createOpenUrlModal} = require('./openUrlModal');
 
 let mainWindow = null;
