+++ src/openUrlModal/view.html
@@ -2,7 +2,7 @@
 <head>
     <title>Open URL...</title>
     <!--<link rel="stylesheet" href="browser.css"/>-->
-    <script src="openUrl.js"></script>
+    <script src="view.js"></script>
 </head>
 <body>
     <p>
